Add files via upload
Firstly, the code imports the necessary modules - numpy and matplotlib.pyplot.

Then, the code defines some constants - Boltzmann constant (k), temperature (T), electron charge (e), Planck constant (h), electron mass (m), and Fermi energy (Ef).

Next, the energy range is defined using the linspace function of numpy, which creates a vector of 1001 equally spaced points between -0.5 and 0.5 eV.

The Fermi-Dirac distribution is calculated using the formula F = 1/(1 + exp((E-Ef)/(k*T/e))), where E is the energy range vector and F is the Fermi-Dirac distribution.

The plot function of matplotlib is then used to plot the Fermi-Dirac distribution as a function of energy E. The title, xlabel, and ylabel functions are used to add appropriate labels to the plot, and the ylim function is used to set the y-axis limits to be between 0 and 1.

Next, the band parameters are defined - conduction band energy (Ec), valence band energy (Ev), and band gap energy (Eg).

The density of states in the conduction and valence bands is calculated using the effective mass approximation. The effective density of states in conduction band Nc is calculated as 2*((2np.pimkT)/(h2))(3/2)np.exp(-(Ec-Ef)/(kT)) and the effective density of states in valence band Nv is calculated as 2*((2np.pimkT)/(h2))(3/2)np.exp(-(Ef-Ev)/(kT)).

Finally, the carrier concentrations (n and p) are calculated using the density of states and the Fermi energy, and the energy band diagram is plotted using the plot function. The legend function adds a legend to the plot indicating the conduction band, valence band, and Fermi level.

The show function is used to display the plots.